refactor(site): rename `<404Page />` to `<NotFoundPage />`
